feat: add cash exchange
