+++ src/main/webapp/resources/js/page/employee-update.js
@@ -5,6 +5,23 @@ $(document).ready(function() {
 
     registerClickEventOfEmpSearchBtn();
     registerClickEventOfEmpUpdateCompleteBtn();
+
+        $('#search-modal-select-btn').on('click', function () {
+
+        const selectedEmployee = $('input[name="selected-employee"]:checked');
+
+        if (selectedEmployee.length > 0) {
+
+            const selectedRow = selectedEmployee.closest('tr');
+            const customerId = selectedRow.find('td:nth-child(2)').text();
+
+            window.location.href = `/page/manager/employee-update?id=${customerId}`;
+
+
+        } else {
+            alert('고객을 선택해 주세요.');
+        }
+    });
 });
